+++ docs/tr2/CHANGELOG.md
@@ -10,6 +10,7 @@
 - added `enable_debug_pos` setting that shows Lara's position in realtime (reachable via `/debug`)
 - added graphics effects to the savegame so they now persist on load (#2736)
 - added an option to control whether or not Lara responds to hitting a wall while wading (#3138)
+- added an option to fix the breakable floor descending glitch (#3152)
 - changed the music track slot limit from 64 to 1024 (#3101)
 - ⚠️ changed the music track behaviour to no longer shift track numbers (#3100)
   - if playing original levels, make sure to update the game flow and injection files from this release